Refactor proxy action in Aave and Spark services
The Aave and Spark services have been updated to use proxy action ABI for the borrowing function, replacing the previous direct interaction with Aave and Spark pool contracts. Additionally, the Aave and Spark address maps now include the proxy action contract address. This is a step towards abstracting actions into proxy contracts which result in a centralized location for logic, simplifying interactions with the Aave and Spark protocols, and improving the code maintainability.